Splitting service portion of restconfig is service specific modules
