+++ src/restconfig-wcs/src/main/java/org/geoserver/rest/service/WCSSettingsController.java
