+++ src/restconfig-wcs/src/main/java/org/geoserver/rest/service/ftl-templates/wcsSettings.ftl
