+++ src/restconfig-wcs/src/test/java/org/geoserver/rest/service/WCSSettingsControllerTest.java
