+++ src/restconfig-wfs/src/main/java/org/geoserver/rest/service/WFSSettingsController.java
