+++ src/restconfig-wfs/src/main/java/org/geoserver/rest/service/ftl-templates/wfsSettings.ftl
