+++ src/restconfig-wms/src/main/java/org/geoserver/rest/service/WMSSettingsController.java
