+++ src/restconfig-wms/src/main/java/org/geoserver/rest/service/ftl-templates/wmsSettings.ftl
