+++ src/restconfig-wms/src/test/java/org/geoserver/rest/service/WMSSettingsControllerTest.java
