+++ src/restconfig-wms/src/test/resources/org/geoserver/rest/service/wcs.xml
@@ -1,0 +1,40 @@
+<wcs>
+  <id>wcs</id>
+  <enabled>true</enabled>
+  <name>WCS</name>
+  <title>Web Coverage Service</title>
+  <maintainer>http://jira.codehaus.org/secure/BrowseProject.jspa?id=10311</maintainer>
+  <abstrct>This server implements the WCS specification 1.0 and 1.1.1, it&apos;s reference implementation of WCS 1.1.1. All layers published by this service are available on WMS also.</abstrct>
+  <accessConstraints>NONE</accessConstraints>
+  <fees>NONE</fees>
+  <versions>
+    <org.geotools.util.Version>
+      <version>1.0.0</version>
+    </org.geotools.util.Version>
+    <org.geotools.util.Version>
+      <version>1.1.1</version>
+    </org.geotools.util.Version>
+    <org.geotools.util.Version>
+      <version>2.0.1</version>
+    </org.geotools.util.Version>
+  </versions>
+  <keywords>
+    <string>WCS</string>
+    <string>WMS</string>
+    <string>GEOSERVER</string>
+  </keywords>
+  <metadataLink>
+    <type>undef</type>
+    <about>http://geoserver.sourceforge.net/html/index.php</about>
+    <metadataType>other</metadataType>
+  </metadataLink>
+  <citeCompliant>false</citeCompliant>
+  <onlineResource>http://geoserver.sourceforge.net/html/index.php</onlineResource>
+  <schemaBaseURL>http://schemas.opengis.net</schemaBaseURL>
+  <verbose>false</verbose>
+  <gmlPrefixing>false</gmlPrefixing>
+  <latLon>false</latLon>
+  <maxInputMemory>0</maxInputMemory>
+  <maxOutputMemory>0</maxOutputMemory>
+  <overviewPolicy>NEAREST</overviewPolicy>
+</wcs>
