+++ src/restconfig/src/test/java/org/geoserver/rest/catalog/FeatureTypeControllerTest.java
@@ -222,77 +222,6 @@ public class FeatureTypeControllerTest extends CatalogRESTTestSupport {
     }
 
     @Test
-    public void testPostAsXML() throws Exception {
-        Document dom = getAsDOM("wfs?request=getfeature&typename=sf:pdsa");
-        assertEquals("ows:ExceptionReport", dom.getDocumentElement().getNodeName());
-
-        addPropertyDataStore(false);
-        String xml = "<featureType>" + "<name>pdsa</name>" + "<nativeName>pdsa</nativeName>"
-                + "<srs>EPSG:4326</srs>" + "<nativeCRS>EPSG:4326</nativeCRS>"
-                + "<nativeBoundingBox>" + "<minx>0.0</minx>" + "<maxx>1.0</maxx>"
-                + "<miny>0.0</miny>" + "<maxy>1.0</maxy>" + "<crs>EPSG:4326</crs>"
-                + "</nativeBoundingBox>" + "<store>pds</store>" + "</featureType>";
-        MockHttpServletResponse response = postAsServletResponse(
-                BASEPATH + "/workspaces/gs/datastores/pds/featuretypes/", xml, "text/xml");
-
-        assertEquals(201, response.getStatus());
-        assertNotNull(response.getHeader("Location"));
-        assertTrue(response.getHeader("Location")
-                .endsWith("/workspaces/gs/datastores/pds/featuretypes/pdsa"));
-
-        dom = getAsDOM("wfs?request=getfeature&typename=gs:pdsa");
-        assertEquals("wfs:FeatureCollection", dom.getDocumentElement().getNodeName());
-        assertEquals(2, dom.getElementsByTagName("gs:pdsa").getLength());
-    }
-
-    @Test
-    public void testPostAsXMLInlineStore() throws Exception {
-        Document dom = getAsDOM("wfs?request=getfeature&typename=sf:pdsa");
-        assertEquals("ows:ExceptionReport", dom.getDocumentElement().getNodeName());
-
-        addPropertyDataStore(false);
-        String xml = "<featureType>" + "<name>pdsa</name>" + "<nativeName>pdsa</nativeName>"
-                + "<srs>EPSG:4326</srs>" + "<nativeCRS>EPSG:4326</nativeCRS>"
-                + "<nativeBoundingBox>" + "<minx>0.0</minx>" + "<maxx>1.0</maxx>"
-                + "<miny>0.0</miny>" + "<maxy>1.0</maxy>" + "<crs>EPSG:4326</crs>"
-                + "</nativeBoundingBox>" + "<store>pds</store>" + "</featureType>";
-        MockHttpServletResponse response = postAsServletResponse(
-                BASEPATH + "/workspaces/gs/featuretypes/", xml, "text/xml");
-
-        assertEquals(201, response.getStatus());
-        assertNotNull(response.getHeader("Location"));
-        assertTrue(response.getHeader("Location")
-                .endsWith("/workspaces/gs/featuretypes/pdsa"));
-
-        dom = getAsDOM("wfs?request=getfeature&typename=gs:pdsa");
-        assertEquals("wfs:FeatureCollection", dom.getDocumentElement().getNodeName());
-        assertEquals(2, dom.getElementsByTagName("gs:pdsa").getLength());
-    }
-
-    @Test
-    public void testPostAsJSON() throws Exception {
-        Document dom = getAsDOM("wfs?request=getfeature&typename=sf:pdsa");
-        assertEquals("ows:ExceptionReport", dom.getDocumentElement().getNodeName());
-
-        addPropertyDataStore(false);
-        String json = "{" + "'featureType':{" + "'name':'pdsa'," + "'nativeName':'pdsa',"
-                + "'srs':'EPSG:4326'," + "'nativeBoundingBox':{" + "'minx':0.0," + "'maxx':1.0,"
-                + "'miny':0.0," + "'maxy':1.0," + "'crs':'EPSG:4326'" + "},"
-                + "'nativeCRS':'EPSG:4326'," + "'store':'pds'" + "}" + "}";
-        MockHttpServletResponse response = postAsServletResponse(
-                BASEPATH + "/workspaces/gs/datastores/pds/featuretypes/", json, "text/json");
-
-        assertEquals(201, response.getStatus());
-        assertNotNull(response.getHeader("Location"));
-        assertTrue(response.getHeader("Location")
-                .endsWith("/workspaces/gs/datastores/pds/featuretypes/pdsa"));
-
-        dom = getAsDOM("wfs?request=getfeature&typename=gs:pdsa");
-        assertEquals("wfs:FeatureCollection", dom.getDocumentElement().getNodeName());
-        assertEquals(2, dom.getElementsByTagName("gs:pdsa").getLength());
-    }
-
-    @Test
     public void testPostToResource() throws Exception {
         addPropertyDataStore(true);
         String xml = "<featureType>" + "<name>pdsa</name>" + "</featureType>";
